+++ lib/ui/widgets/triangle_header_image.dart
@@ -1,0 +1,52 @@
+import 'package:flutter/cupertino.dart';
+import 'package:rokwire_plugin/service/config.dart';
+import 'package:rokwire_plugin/service/styles.dart';
+import 'package:rokwire_plugin/ui/panels/modal_image_holder.dart';
+import 'package:rokwire_plugin/ui/widgets/triangle_painter.dart';
+
+class TriangleHeaderImage extends StatelessWidget {
+  final Color?  flexBackColor;
+  final String? flexImageUrl;
+  final String? flexImageKey;
+  final Color?  flexLeftToRightTriangleColor;
+  final double? flexLeftToRightTriangleHeight;
+  final Color?  flexRightToLeftTriangleColor;
+  final double? flexRightToLeftTriangleHeight;
+  const TriangleHeaderImage({super.key, this.flexBackColor, this.flexImageUrl, this.flexImageKey, this.flexLeftToRightTriangleColor,
+    this.flexLeftToRightTriangleHeight, this.flexRightToLeftTriangleColor, this.flexRightToLeftTriangleHeight
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(color: flexBackColor, child:
+      Stack(alignment: Alignment.bottomCenter, children: <Widget>[
+        buildFlexibleInterior(context),
+        buildFlexibleLeftToRightTriangle(context),
+        buildFlexibleLeftTriangle(context),
+      ],),
+    );
+  }
+
+  @protected
+  Widget buildFlexibleInterior(BuildContext context) {
+    Widget? image;
+    if (flexImageUrl != null) {
+      image = Image.network(flexImageUrl!, fit: BoxFit.cover, headers: Config().networkAuthHeaders, excludeFromSemantics: true);
+    } else if (flexImageKey != null) {
+      image = Styles().images.getImage(flexImageKey, fit: BoxFit.cover, excludeFromSemantics: true);
+    }
+    return (image != null) ? Positioned.fill(child: ModalImageHolder(child:image)) : Container();
+  }
+
+  @protected
+  Widget buildFlexibleLeftToRightTriangle(BuildContext context) => CustomPaint(
+    painter: TrianglePainter(painterColor: flexLeftToRightTriangleColor, horzDir: TriangleHorzDirection.leftToRight),
+    child: Container(height: flexLeftToRightTriangleHeight,),
+  );
+
+  @protected
+  Widget buildFlexibleLeftTriangle(BuildContext context) => CustomPaint(
+    painter: TrianglePainter(painterColor: flexRightToLeftTriangleColor),
+    child: Container(height: flexRightToLeftTriangleHeight,),
+  );
+}